Add files via upload
created 'The Cake Delight,' a visually beautiful and responsive HTML and CSS website for a bakery. By offering an easy-to-use interface, the project aims to improve the bakery's online visibility. A comprehensive product inventory, an interactive gallery, and an intuitive layout are among the salient features. shown mastery in responsive design methodologies, layout organization, and web design principles